requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
google-auth==2.29.0
firebase-admin==6.5.0
Flask==3.0.3